+++ src/error.rs
@@ -1,6 +1,7 @@
 use std::num::TryFromIntError;
 
-use cosmwasm_std::StdError;
+use cosmwasm_std::{Response, StdError};
+use provwasm_std::ProvenanceMsg;
 use thiserror::Error;
 
 #[derive(Error, Debug)]
@@ -26,6 +27,6 @@ impl From<TryFromIntError> for ContractError {
     }
 }
 
-pub fn contract_error<T>(err: &str) -> Result<T, ContractError> {
+pub fn contract_error(err: &str) -> Result<Response<ProvenanceMsg>, ContractError> {
     Err(ContractError::Std(StdError::generic_err(err)))
 }
