+++ src/subscribe.rs
@@ -6,13 +6,10 @@ use crate::state::{asset_exchange_storage, AssetExchange};
 use crate::sub_msg::{SubInstantiateMsg, SubQueryMsg, SubState};
 use cosmwasm_std::Deps;
 use cosmwasm_std::DepsMut;
-use cosmwasm_std::Env;
 use cosmwasm_std::MessageInfo;
 use cosmwasm_std::Response;
 use cosmwasm_std::StdResult;
-use cosmwasm_std::SubMsg;
-use cosmwasm_std::WasmMsg;
-use cosmwasm_std::{to_binary, Addr};
+use cosmwasm_std::{to_binary, Addr, Env, SubMsg, WasmMsg};
 use provwasm_std::ProvenanceQuerier;
 use provwasm_std::ProvenanceQuery;
 use std::collections::HashSet;
@@ -183,13 +180,13 @@ mod tests {
     use crate::msg::QueryMsg;
     use crate::msg::RaiseState;
     use crate::query::query;
+    use crate::state::accepted_subscriptions_read;
     use crate::state::config;
     use crate::state::pending_subscriptions_read;
-    use crate::state::tests::set_accepted;
     use crate::state::tests::set_pending;
     use crate::state::tests::to_addresses;
+    use crate::state::tests::{asset_exchange_storage_read, set_accepted};
     use crate::state::State;
-    use crate::state::{accepted_subscriptions_read, asset_exchange_storage_read};
     use cosmwasm_std::coins;
     use cosmwasm_std::from_binary;
     use cosmwasm_std::testing::mock_env;
@@ -364,93 +361,6 @@ mod tests {
         assert!(res.is_err());
     }
 
-    // #[test]
-    // fn close_remaining_commitment() {
-    //     let mut deps = default_deps(None);
-    //     load_markers(&mut deps.querier);
-    //     outstanding_subscription_closures(&mut deps.storage)
-    //         .save(&to_addresses(vec!["sub_1"]))
-    //         .unwrap();
-
-    //     // close remaining commitment
-    //     let res = execute(
-    //         deps.as_mut(),
-    //         mock_env(),
-    //         mock_info("sub_1", &coins(100, "commitment_coin")),
-    //         HandleMsg::CloseRemainingCommitment {},
-    //     )
-    //     .unwrap();
-
-    //     assert_eq!(2, res.messages.len());
-
-    //     // verify deposit commitment
-    //     let (to_address, coins) = send_args(msg_at_index(&res, 0));
-    //     assert_eq!("tp18vmzryrvwaeykmdtu6cfrz5sau3dhc5c73ms0u", to_address);
-    //     assert_eq!("commitment_coin", coins.first().unwrap().denom);
-    //     assert_eq!(100, coins.first().unwrap().amount.u128());
-
-    //     // verify burn commitment
-    //     let coin = burn_args(msg_at_index(&res, 1));
-    //     assert_eq!("commitment_coin", coin.denom);
-    //     assert_eq!(100, coin.amount.u128());
-
-    //     // verify outsounding sub closure removed
-    //     assert_eq!(
-    //         0,
-    //         outstanding_subscription_closures(&mut deps.storage)
-    //             .load()
-    //             .unwrap()
-    //             .len()
-    //     );
-    //     // verify closed sub exists
-    //     assert_eq!(
-    //         1,
-    //         closed_subscriptions(&mut deps.storage)
-    //             .load()
-    //             .unwrap()
-    //             .len()
-    //     )
-    // }
-
-    // #[test]
-    // fn close_remaining_commitment_not_found() {
-    //     let mut deps = default_deps(None);
-    //     outstanding_subscription_closures(&mut deps.storage)
-    //         .save(&to_addresses(vec!["sub_1"]))
-    //         .unwrap();
-
-    //     // close remaining commitment
-    //     let res = execute(
-    //         deps.as_mut(),
-    //         mock_env(),
-    //         mock_info("sub_2", &coins(100, "commitment_coin")),
-    //         HandleMsg::CloseRemainingCommitment {},
-    //     );
-
-    //     assert!(res.is_err());
-    // }
-
-    // #[test]
-    // fn close_remaining_commitment_leftover() {
-    //     let mut deps = default_deps(None);
-    //     outstanding_subscription_closures(&mut deps.storage)
-    //         .save(&to_addresses(vec!["sub_1"]))
-    //         .unwrap();
-    //     deps.querier
-    //         .base
-    //         .update_balance(Addr::unchecked("sub_1"), coins(100, "commitment_coin"));
-
-    //     // close remaining commitment
-    //     let res = execute(
-    //         deps.as_mut(),
-    //         mock_env(),
-    //         mock_info("sub_1", &coins(100, "commitment_coin")),
-    //         HandleMsg::CloseRemainingCommitment {},
-    //     );
-
-    //     assert!(res.is_err());
-    // }
-
     #[test]
     fn accept_subscription() {
         let mut deps = mock_sub_state();
@@ -588,368 +498,4 @@ mod tests {
         );
         assert!(res.is_err());
     }
-
-    // #[test]
-    // fn update_commitments_increase() {
-    //     let mut deps = default_deps(None);
-    //     set_accepted(&mut deps.storage, vec!["sub_1"]);
-    //     deps.querier
-    //         .base
-    //         .update_balance(Addr::unchecked("sub_1"), coins(1_000, "commitment_coin"));
-
-    //     // update commitments
-    //     execute(
-    //         deps.as_mut(),
-    //         mock_env(),
-    //         mock_info("gp", &[]),
-    //         HandleMsg::UpdateCommitments {
-    //             commitment_updates: vec![CommitmentUpdate {
-    //                 subscription: Addr::unchecked("sub_1"),
-    //                 change_by_amount: 1_000,
-    //             }]
-    //             .into_iter()
-    //             .collect(),
-    //         },
-    //     )
-    //     .unwrap();
-
-    //     // verify outsounding commitment update exists
-    //     assert_eq!(
-    //         1,
-    //         outstanding_commitment_updates(&mut deps.storage)
-    //             .load()
-    //             .unwrap()
-    //             .len()
-    //     )
-    // }
-
-    // #[test]
-    // fn update_commitments_decrease() {
-    //     let mut deps = default_deps(None);
-    //     set_accepted(&mut deps.storage, vec!["sub_1"]);
-    //     deps.querier
-    //         .base
-    //         .update_balance(Addr::unchecked("sub_1"), coins(1_000, "commitment_coin"));
-
-    //     // update commitments
-    //     execute(
-    //         deps.as_mut(),
-    //         mock_env(),
-    //         mock_info("gp", &[]),
-    //         HandleMsg::UpdateCommitments {
-    //             commitment_updates: vec![CommitmentUpdate {
-    //                 subscription: Addr::unchecked("sub_1"),
-    //                 change_by_amount: -1_000,
-    //             }]
-    //             .into_iter()
-    //             .collect(),
-    //         },
-    //     )
-    //     .unwrap();
-
-    //     // verify outsounding commitment update exists
-    //     assert_eq!(
-    //         1,
-    //         outstanding_commitment_updates(&mut deps.storage)
-    //             .load()
-    //             .unwrap()
-    //             .len()
-    //     )
-    // }
-
-    // #[test]
-    // fn update_commitments_stomp() {
-    //     let mut deps = default_deps(None);
-    //     set_accepted(&mut deps.storage, vec!["sub_1"]);
-    //     outstanding_commitment_updates(&mut deps.storage)
-    //         .save(
-    //             &vec![CommitmentUpdate {
-    //                 subscription: Addr::unchecked("sub_1"),
-    //                 change_by_amount: 1_000,
-    //             }]
-    //             .into_iter()
-    //             .collect(),
-    //         )
-    //         .unwrap();
-    //     deps.querier
-    //         .base
-    //         .update_balance(Addr::unchecked("sub_1"), coins(1_000, "commitment_coin"));
-
-    //     // update commitments
-    //     execute(
-    //         deps.as_mut(),
-    //         mock_env(),
-    //         mock_info("gp", &[]),
-    //         HandleMsg::UpdateCommitments {
-    //             commitment_updates: vec![CommitmentUpdate {
-    //                 subscription: Addr::unchecked("sub_1"),
-    //                 change_by_amount: 2_000,
-    //             }]
-    //             .into_iter()
-    //             .collect(),
-    //         },
-    //     )
-    //     .unwrap();
-
-    //     // verify outsounding commitment update exists
-    //     let updates = outstanding_commitment_updates(&mut deps.storage)
-    //         .load()
-    //         .unwrap();
-    //     assert_eq!(1, updates.len());
-    //     assert_eq!(
-    //         2_000,
-    //         Vec::from_iter(updates).get(0).unwrap().change_by_amount
-    //     );
-    // }
-
-    // #[test]
-    // fn update_commitments_decrease_too_much() {
-    //     let mut deps = default_deps(None);
-    //     set_pending(&mut deps.storage, vec!["sub_1"]);
-    //     deps.querier
-    //         .base
-    //         .update_balance(Addr::unchecked("sub_1"), coins(1_000, "commitment_coin"));
-
-    //     // update commitments
-    //     let res = execute(
-    //         deps.as_mut(),
-    //         mock_env(),
-    //         mock_info("gp", &[]),
-    //         HandleMsg::UpdateCommitments {
-    //             commitment_updates: vec![CommitmentUpdate {
-    //                 subscription: Addr::unchecked("sub_1"),
-    //                 change_by_amount: -2_000,
-    //             }]
-    //             .into_iter()
-    //             .collect(),
-    //         },
-    //     );
-
-    //     assert!(res.is_err());
-    // }
-
-    // #[test]
-    // fn update_commitments_decrease_zero() {
-    //     let mut deps = default_deps(None);
-    //     set_accepted(&mut deps.storage, vec!["sub_1"]);
-    //     deps.querier
-    //         .base
-    //         .update_balance(Addr::unchecked("sub_1"), coins(1_000, "commitment_coin"));
-
-    //     // update commitments
-    //     let res = execute(
-    //         deps.as_mut(),
-    //         mock_env(),
-    //         mock_info("gp", &[]),
-    //         HandleMsg::UpdateCommitments {
-    //             commitment_updates: vec![CommitmentUpdate {
-    //                 subscription: Addr::unchecked("sub_1"),
-    //                 change_by_amount: 0,
-    //             }]
-    //             .into_iter()
-    //             .collect(),
-    //         },
-    //     );
-
-    //     assert!(res.is_err());
-    // }
-
-    // #[test]
-    // fn update_commitments_bad_actor() {
-    //     let mut deps = default_deps(None);
-    //     set_accepted(&mut deps.storage, vec!["sub_1"]);
-    //     deps.querier
-    //         .base
-    //         .update_balance(Addr::unchecked("sub_1"), coins(1_000, "commitment_coin"));
-
-    //     // update commitments
-    //     let res = execute(
-    //         deps.as_mut(),
-    //         mock_env(),
-    //         mock_info("bad_actor", &[]),
-    //         HandleMsg::UpdateCommitments {
-    //             commitment_updates: vec![CommitmentUpdate {
-    //                 subscription: Addr::unchecked("sub_1"),
-    //                 change_by_amount: 1_000,
-    //             }]
-    //             .into_iter()
-    //             .collect(),
-    //         },
-    //     );
-
-    //     assert!(res.is_err());
-    // }
-
-    // #[test]
-    // fn accept_commitment_update_increase() {
-    //     let mut deps = default_deps(None);
-    //     load_markers(&mut deps.querier);
-    //     outstanding_commitment_updates(&mut deps.storage)
-    //         .save(
-    //             &vec![CommitmentUpdate {
-    //                 subscription: Addr::unchecked("sub_1"),
-    //                 change_by_amount: 1_000,
-    //             }]
-    //             .into_iter()
-    //             .collect(),
-    //         )
-    //         .unwrap();
-
-    //     // accept commitment update
-    //     let res = execute(
-    //         deps.as_mut(),
-    //         mock_env(),
-    //         mock_info("sub_1", &vec![]),
-    //         HandleMsg::AcceptCommitmentUpdate {},
-    //     )
-    //     .unwrap();
-
-    //     assert_eq!(2, res.messages.len());
-
-    //     // verify minted coin
-    //     let mint = mint_args(msg_at_index(&res, 0));
-    //     assert_eq!(1_000, mint.amount.u128());
-    //     assert_eq!("commitment_coin", mint.denom);
-
-    //     // verify withdrawn coin
-    //     let (marker_denom, coin, recipient) = withdraw_args(msg_at_index(&res, 1));
-    //     assert_eq!("commitment_coin", marker_denom);
-    //     assert_eq!(1_000, coin.amount.u128());
-    //     assert_eq!("commitment_coin", coin.denom);
-    //     assert_eq!("sub_1", recipient.clone().into_string());
-
-    //     // verify outsounding commitment update removed
-    //     assert_eq!(
-    //         0,
-    //         outstanding_commitment_updates(&mut deps.storage)
-    //             .load()
-    //             .unwrap()
-    //             .len()
-    //     );
-    // }
-
-    // #[test]
-    // fn accept_commitment_update_decrease() {
-    //     let mut deps = default_deps(None);
-    //     load_markers(&mut deps.querier);
-    //     outstanding_commitment_updates(&mut deps.storage)
-    //         .save(
-    //             &vec![CommitmentUpdate {
-    //                 subscription: Addr::unchecked("sub_1"),
-    //                 change_by_amount: -1_000,
-    //             }]
-    //             .into_iter()
-    //             .collect(),
-    //         )
-    //         .unwrap();
-
-    //     // accept commitment update
-    //     let res = execute(
-    //         deps.as_mut(),
-    //         mock_env(),
-    //         mock_info("sub_1", &coins(1_000, "commitment_coin")),
-    //         HandleMsg::AcceptCommitmentUpdate {},
-    //     )
-    //     .unwrap();
-
-    //     assert_eq!(2, res.messages.len());
-
-    //     // verify deposit commitment
-    //     let (to_address, coins) = send_args(msg_at_index(&res, 0));
-    //     assert_eq!("tp18vmzryrvwaeykmdtu6cfrz5sau3dhc5c73ms0u", to_address);
-    //     assert_eq!("commitment_coin", coins.first().unwrap().denom);
-    //     assert_eq!(1_000, coins.first().unwrap().amount.u128());
-
-    //     // verify burn commitment
-    //     let coin = burn_args(msg_at_index(&res, 1));
-    //     assert_eq!("commitment_coin", coin.denom);
-    //     assert_eq!(1_000, coin.amount.u128());
-
-    //     // verify outsounding commitment update removed
-    //     assert_eq!(
-    //         0,
-    //         outstanding_commitment_updates(&mut deps.storage)
-    //             .load()
-    //             .unwrap()
-    //             .len()
-    //     );
-    // }
-
-    // #[test]
-    // fn accept_commitment_update_decrease_no_funds() {
-    //     let mut deps = default_deps(None);
-    //     load_markers(&mut deps.querier);
-    //     outstanding_commitment_updates(&mut deps.storage)
-    //         .save(
-    //             &vec![CommitmentUpdate {
-    //                 subscription: Addr::unchecked("sub_1"),
-    //                 change_by_amount: -1_000,
-    //             }]
-    //             .into_iter()
-    //             .collect(),
-    //         )
-    //         .unwrap();
-
-    //     // accept commitment update
-    //     let res = execute(
-    //         deps.as_mut(),
-    //         mock_env(),
-    //         mock_info("sub_1", &vec![]),
-    //         HandleMsg::AcceptCommitmentUpdate {},
-    //     );
-
-    //     assert!(res.is_err())
-    // }
-
-    // #[test]
-    // fn accept_commitment_update_decrease_bad_denom() {
-    //     let mut deps = default_deps(None);
-    //     load_markers(&mut deps.querier);
-    //     outstanding_commitment_updates(&mut deps.storage)
-    //         .save(
-    //             &vec![CommitmentUpdate {
-    //                 subscription: Addr::unchecked("sub_1"),
-    //                 change_by_amount: -1_000,
-    //             }]
-    //             .into_iter()
-    //             .collect(),
-    //         )
-    //         .unwrap();
-
-    //     // accept commitment update
-    //     let res = execute(
-    //         deps.as_mut(),
-    //         mock_env(),
-    //         mock_info("sub_1", &coins(1_000, "bad_denom")),
-    //         HandleMsg::AcceptCommitmentUpdate {},
-    //     );
-
-    //     assert!(res.is_err())
-    // }
-
-    // #[test]
-    // fn accept_commitment_update_decrease_bad_amount() {
-    //     let mut deps = default_deps(None);
-    //     load_markers(&mut deps.querier);
-    //     outstanding_commitment_updates(&mut deps.storage)
-    //         .save(
-    //             &vec![CommitmentUpdate {
-    //                 subscription: Addr::unchecked("sub_1"),
-    //                 change_by_amount: -1_000,
-    //             }]
-    //             .into_iter()
-    //             .collect(),
-    //         )
-    //         .unwrap();
-
-    //     // accept commitment update
-    //     let res = execute(
-    //         deps.as_mut(),
-    //         mock_env(),
-    //         mock_info("sub_1", &coins(100, "commitment_denom")),
-    //         HandleMsg::AcceptCommitmentUpdate {},
-    //     );
-
-    //     assert!(res.is_err())
-    // }
 }
